Extract TextRenderSystem
