+++ src/systems/mod.rs
@@ -1,8 +1,6 @@
 use std::time::Duration;
-use resources::CanvasHolder;
-use builders::{ TextBuilder, TextTexture };
-use specs::{ System, Read, Write, ReadStorage, WriteStorage };
-use components::{ Position, Text, FPS };
+use specs::{ System, Read, WriteStorage };
+use components::{ Text, FPS };
 use resources::{ DeltaTime };
 
 mod update_position;
@@ -11,41 +9,8 @@ pub use self::update_position::UpdatePos;
 mod draw_system;
 pub use self::draw_system::DrawSystem;
 
-pub struct TextRenderSystem<'b> {
-    text_builder: TextBuilder<'b>
-}
-
-impl<'b> TextRenderSystem<'b> {
-    pub fn new(text_builder: TextBuilder<'b>) -> Self {
-        TextRenderSystem {
-            text_builder: text_builder
-        }
-    }
-}
-
-impl<'a, 'b> System<'a> for TextRenderSystem<'b> {
-    type SystemData = (
-        Write<'a, CanvasHolder>,
-        ReadStorage<'a, Text>,
-        ReadStorage<'a, Position>
-    );
-
-    fn run<'c>(&'c mut self, data: Self::SystemData) {
-        use specs::Join;
-        use sdl2::rect::Rect;
-
-        let (mut canvas_holder, text, pos) = data;
-
-        for (text, pos) in (&text, &pos).join() {
-            let Text { text, offset, color, font } = text;
-            let text_texture: TextTexture<'c> = self.text_builder.build_text(text, font, color);
-            let message_target = Rect::new(pos.x as i32 + offset.x, pos.y as i32 + offset.y, text_texture.query.width, text_texture.query.height);
-
-            let texture = text_texture.texture;
-            canvas_holder.borrow().unwrap().copy(&texture, None, Some(message_target)).expect("could not copy texture to canvas");
-        }
-    }
-}
+mod text_render_system;
+pub use self::text_render_system::TextRenderSystem;
 
 pub struct FpsCounter {
     counter: u16,
